+++ example/lib/utils/settings_controller.dart
@@ -6,7 +6,6 @@ class ExampleSettings {
   final Brightness brightness;
   final bool scrollable;
   final bool showFooter;
-  final Color? backgroundColor;
 
   // Provider-specific params
   final VimeoEmbedParams vimeoParams;
@@ -21,7 +20,6 @@ class ExampleSettings {
     this.brightness = Brightness.light,
     this.scrollable = false,
     this.showFooter = false,
-    this.backgroundColor,
     this.vimeoParams = const VimeoEmbedParams(),
     this.xParams = const XEmbedParams(
       dnt: true,
@@ -42,7 +40,6 @@ class ExampleSettings {
     Brightness? brightness,
     bool? scrollable,
     bool? showFooter,
-    Color? backgroundColor,
     VimeoEmbedParams? vimeoParams,
     XEmbedParams? xParams,
     MetaEmbedParams? metaParams,
@@ -55,7 +52,6 @@ class ExampleSettings {
       brightness: brightness ?? this.brightness,
       scrollable: scrollable ?? this.scrollable,
       showFooter: showFooter ?? this.showFooter,
-      backgroundColor: backgroundColor ?? this.backgroundColor,
       vimeoParams: vimeoParams ?? this.vimeoParams,
       xParams: xParams ?? this.xParams,
       metaParams: metaParams ?? this.metaParams,
@@ -91,14 +87,12 @@ class ExampleSettingsController extends ChangeNotifier {
     Brightness? brightness,
     bool? scrollable,
     bool? showFooter,
-    Color? backgroundColor,
   }) {
     _settings = _settings.copyWith(
       locale: locale,
       brightness: brightness,
       scrollable: scrollable,
       showFooter: showFooter,
-      backgroundColor: backgroundColor,
     );
     notifyListeners();
   }