+++ lib/src/core/embed_scope.dart
@@ -1,10 +1,10 @@
+import 'package:flutter_oembed/src/cache/in_memory_cache_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_oembed/src/core/embed_cache_provider.dart';
 import 'package:flutter_oembed/src/models/params/base_embed_params.dart';
 import 'package:flutter_oembed/src/models/configs/embed_config.dart';
 import 'package:flutter_oembed/src/models/core/embed_enums.dart';
 import 'package:flutter_oembed/src/models/core/embed_style.dart';
-import 'package:flutter_oembed/src/services/default_embed_cache_provider.dart';
 import 'package:flutter_oembed/src/services/embed_service.dart';
 
 /// Provides [EmbedConfig] to the widget subtree.
@@ -107,7 +107,7 @@ class EmbedScope extends InheritedWidget {
   }) {
     return cacheProvider ??
         config?.cacheProvider ??
-        DefaultEmbedCacheProvider.instance;
+        InMemoryEmbedCacheProvider.instance;
   }
 
   @override